refactored avatar page into slotrow, saveoutfitmodal, and slots components
