+++ client/app/(app)/(tabs)/avatar.tsx
@@ -1,121 +1,19 @@
-import MaterialIcons from '@expo/vector-icons/MaterialIcons';
-import { StatusBar } from 'expo-status-bar';
-import React, { useEffect, useMemo, useRef, useState } from 'react';
-import {
-  ActivityIndicator,
-  Modal,
-  PanResponder,
-  Pressable,
-  StyleSheet,
-  Text,
-  TextInput,
-  View,
-} from 'react-native';
 import { useBottomTabBarHeight } from '@react-navigation/bottom-tabs';
+import { StatusBar } from 'expo-status-bar';
+import { useEffect, useMemo, useState } from 'react';
+import { ActivityIndicator, Pressable, StyleSheet, Text, View } from 'react-native';
+import MaterialIcons from '@expo/vector-icons/MaterialIcons';
 
-import { Image } from 'expo-image';
+import { SaveOutfitModal } from '@/components/avatar/save-outfit-modal';
+import { SlotRow } from '@/components/avatar/slot-row';
+import { SLOTS, SUB_TO_CAT } from '@/components/avatar/slots';
+import type { SlotCategory, SlotIndices } from '@/components/avatar/slots';
 import { MainHeader } from '@/components/ui/main-header';
-import { Category, SUBCATEGORIES } from '@/constants/categories';
+import { Category } from '@/constants/categories';
 import { Palette, Radius, Spacing, Typography } from '@/constants/design';
 import { fetchClosetItems } from '@/lib/items';
 import { supabase } from '@/lib/supabase';
 import type { ClosetItem } from '@/lib/types/closet';
-
-const SLOTS = [
-  { label: 'Top', category: Category.TOPS, slot: 'top', slotFlex: 6 },
-  { label: 'Bottom', category: Category.BOTTOMS, slot: 'bottom', slotFlex: 7 },
-  { label: 'Shoes', category: Category.SHOES, slot: 'footwear', slotFlex: 5 },
-] as const;
-
-type SlotCategory = (typeof SLOTS)[number]['category'];
-type SlotIndices = Record<SlotCategory, number>;
-const SLOT_CATEGORIES = new Set<string>([Category.TOPS, Category.BOTTOMS, Category.SHOES]);
-const SUB_TO_CAT: Record<string, SlotCategory> = {};
-for (const [cat, subs] of Object.entries(SUBCATEGORIES)) {
-  if (SLOT_CATEGORIES.has(cat)) {
-    for (const sub of subs) {
-      SUB_TO_CAT[sub.toLowerCase()] = cat as SlotCategory;
-    }
-  }
-}
-
-type SlotRowProps = {
-  label: string;
-  category: SlotCategory;
-  slotItems: ClosetItem[];
-  idx: number;
-  onNavigate: (dir: 1 | -1) => void;
-  slotFlex: number;
-};
-
-function SlotRow({ label, slotItems, idx, onNavigate, slotFlex }: SlotRowProps) {
-  const onNavigateRef = useRef(onNavigate);
-  useEffect(() => {
-    onNavigateRef.current = onNavigate;
-  }, [onNavigate]);
-
-  const panResponder = useRef(
-    PanResponder.create({
-      onMoveShouldSetPanResponder: (_, gs) =>
-        Math.abs(gs.dx) > Math.abs(gs.dy) * 1.5 && Math.abs(gs.dx) > 12,
-      onPanResponderRelease: (_, gs) => {
-        if (gs.dx < -50) onNavigateRef.current(1);
-        else if (gs.dx > 50) onNavigateRef.current(-1);
-      },
-    })
-  ).current;
-
-  const current = slotItems[idx] ?? null;
-
-  return (
-    <View style={[styles.slot, { flex: slotFlex }]}>
-      <View style={styles.slotRow}>
-        <Pressable
-          style={[styles.arrowBtn, slotItems.length < 2 && styles.arrowBtnDisabled]}
-          onPress={() => onNavigate(-1)}
-          disabled={slotItems.length < 2}
-        >
-          <MaterialIcons
-            name="chevron-left"
-            size={28}
-            color={slotItems.length < 2 ? Palette.outlineVariant : Palette.onSurface}
-          />
-        </Pressable>
-
-        <View style={styles.itemCard} {...panResponder.panHandlers}>
-          {current?.image_url ? (
-            <Image
-              source={{ uri: current.image_url }}
-              style={styles.itemImage}
-              contentFit="contain"
-            />
-          ) : (
-            <View style={styles.itemPlaceholder}>
-              <Text style={styles.placeholderText}>
-                {slotItems.length === 0
-                  ? `No ${label.toLowerCase()}${label === 'Shoes' ? '' : 's'} in closet`
-                  : 'No image'}
-              </Text>
-            </View>
-          )}
-        </View>
-
-        <Pressable
-          style={[styles.arrowBtn, slotItems.length < 2 && styles.arrowBtnDisabled]}
-          onPress={() => onNavigate(1)}
-          disabled={slotItems.length < 2}
-        >
-          <MaterialIcons
-            name="chevron-right"
-            size={28}
-            color={slotItems.length < 2 ? Palette.outlineVariant : Palette.onSurface}
-          />
-        </Pressable>
-      </View>
-
-    </View>
-  );
-}
 
 export default function AvatarScreen() {
   const tabBarHeight = useBottomTabBarHeight();
@@ -127,7 +25,7 @@ export default function AvatarScreen() {
     [Category.BOTTOMS]: 0,
     [Category.SHOES]: 0,
   });
-  const [saveModalVisible, setSaveModalVisible] = useState(false);
+  const [modalVisible, setModalVisible] = useState(false);
   const [addToFavorites, setAddToFavorites] = useState(false);
   const [outfitName, setOutfitName] = useState('');
   const [saving, setSaving] = useState(false);
@@ -138,14 +36,8 @@ export default function AvatarScreen() {
       try {
         const { data: { session } } = await supabase.auth.getSession();
         const uid = session?.user?.id;
-        if (!uid) {
-          setError('Sign in to create outfits.');
-          setLoading(false);
-          return;
-        }
-        console.log('[Avatar] loading items for uid:', uid);
+        if (!uid) { setError('Sign in to create outfits.'); setLoading(false); return; }
         const fetched = await fetchClosetItems(uid);
-        console.log('[Avatar] fetched items:', fetched.map(i => ({ id: i.id, category: i.category, name: i.item_name })));
         setItems(fetched);
       } catch (e) {
         setError(e instanceof Error ? e.message : 'Failed to load items.');
@@ -163,10 +55,9 @@ export default function AvatarScreen() {
       [Category.SHOES]: [],
     };
     for (const item of items) {
-      // Primary: look up the slot by subcategory (ground truth — comes from a fixed list).
-      // Fallback: use the stored category field, normalised to lowercase.
       const catFromSub = item.subcategory ? SUB_TO_CAT[item.subcategory.toLowerCase()] : undefined;
       const catFromField = item.category?.toLowerCase() as SlotCategory | undefined;
+
       const cat = catFromSub ?? catFromField;
       if (cat && cat in result) result[cat].push(item);
     }
@@ -178,17 +69,14 @@ export default function AvatarScreen() {
   const navigate = (category: SlotCategory, dir: 1 | -1) => {
     const len = grouped[category].length;
     if (len === 0) return;
-    setIndices((prev) => ({
-      ...prev,
-      [category]: (prev[category] + dir + len) % len,
-    }));
+    setIndices((prev) => ({ ...prev, [category]: (prev[category] + dir + len) % len }));
   };
 
-  const openSaveModal = () => {
+  const openModal = () => {
     setAddToFavorites(false);
     setOutfitName('');
     setSaveError(null);
-    setSaveModalVisible(true);
+    setModalVisible(true);
   };
 
   const saveOutfit = async () => {
@@ -217,7 +105,7 @@ export default function AvatarScreen() {
       const { error: itemsError } = await supabase.from('outfit_items').insert(outfitItems);
       if (itemsError) throw itemsError;
 
-      setSaveModalVisible(false);
+      setModalVisible(false);
     } catch (e) {
       setSaveError(e instanceof Error ? e.message : 'Failed to save outfit.');
     } finally {
@@ -242,11 +130,11 @@ export default function AvatarScreen() {
   }
 
   return (
-    <View style={styles.safeArea}>
+    <View style={styles.screen}>
       <StatusBar style="dark" />
       <MainHeader />
 
-      <View style={[styles.screen, { paddingBottom: tabBarHeight }]}>
+      <View style={[styles.body, { paddingBottom: tabBarHeight }]}>
         {SLOTS.map(({ label, category, slotFlex }) => (
           <SlotRow
             key={category}
@@ -262,7 +150,7 @@ export default function AvatarScreen() {
         <View style={styles.actions}>
           <Pressable
             style={[styles.saveBtn, !canSave && styles.saveBtnDisabled]}
-            onPress={openSaveModal}
+            onPress={openModal}
             disabled={!canSave}
           >
             <MaterialIcons name="bookmark-border" size={20} color={Palette.onPrimary} />
@@ -271,72 +159,23 @@ export default function AvatarScreen() {
         </View>
       </View>
 
-      <Modal
-        visible={saveModalVisible}
-        transparent
-        animationType="slide"
-        onRequestClose={() => setSaveModalVisible(false)}
-      >
-        <Pressable style={styles.modalOverlay} onPress={() => setSaveModalVisible(false)}>
-          <Pressable style={styles.modalSheet} onPress={() => {}}>
-            <View style={styles.modalHandle} />
-            <Text style={styles.modalTitle}>Save Outfit</Text>
-
-            <TextInput
-              style={styles.nameInput}
-              placeholder="Outfit name (optional)"
-              placeholderTextColor={Palette.onSurfaceVariant}
-              value={outfitName}
-              onChangeText={setOutfitName}
-              maxLength={80}
-              returnKeyType="done"
-            />
-
-            <Pressable
-              style={styles.favRow}
-              onPress={() => setAddToFavorites((v) => !v)}
-            >
-              <MaterialIcons
-                name={addToFavorites ? 'favorite' : 'favorite-border'}
-                size={24}
-                color={addToFavorites ? Palette.error : Palette.onSurface}
-              />
-              <Text style={styles.favLabel}>Add to Favorites</Text>
-            </Pressable>
-
-            <View style={styles.collectionsRow}>
-              <Text style={styles.collectionsLabel}>Save to Collection</Text>
-              <Text style={styles.comingSoon}>Coming soon</Text>
-            </View>
-
-            {saveError && <Text style={styles.errorText}>{saveError}</Text>}
-
-            <View style={styles.modalActions}>
-              <Pressable
-                style={styles.cancelBtn}
-                onPress={() => setSaveModalVisible(false)}
-              >
-                <Text style={styles.cancelBtnText}>Cancel</Text>
-              </Pressable>
-              <Pressable
-                style={[styles.confirmBtn, saving && styles.saveBtnDisabled]}
-                onPress={saveOutfit}
-                disabled={saving}
-              >
-                <Text style={styles.confirmBtnText}>
-                  {saving ? 'Saving…' : 'Save'}
-                </Text>
-              </Pressable>
-            </View>
-          </Pressable>
-        </Pressable>
-      </Modal>
+      <SaveOutfitModal
+        visible={modalVisible}
+        outfitName={outfitName}
+        addToFavorites={addToFavorites}
+        saving={saving}
+        saveError={saveError}
+        onChangeName={setOutfitName}
+        onToggleFavorites={() => setAddToFavorites((v) => !v)}
+        onSave={saveOutfit}
+        onClose={() => setModalVisible(false)}
+      />
     </View>
   );
 }
 
 const styles = StyleSheet.create({
-  safeArea: {
+  screen: {
     flex: 1,
     backgroundColor: Palette.background,
   },
@@ -346,61 +185,11 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     backgroundColor: Palette.background,
   },
-  screen: {
+  body: {
     flex: 1,
     paddingHorizontal: Spacing.containerMargin,
     paddingTop: Spacing.stackMd,
     gap: Spacing.stackSm,
-  },
-  slot: {
-    // flex is applied inline per slot via slotFlex
-    minHeight: 0,
-    gap: 4,
-  },
-  slotLabel: {
-    ...Typography.labelSm,
-    color: Palette.onSurfaceVariant,
-  },
-  slotRow: {
-    flex: 1,
-    flexDirection: 'row',
-    alignItems: 'stretch',
-    gap: Spacing.stackSm,
-  },
-  arrowBtn: {
-    padding: Spacing.stackSm,
-    borderRadius: Radius.full,
-    backgroundColor: Palette.surfaceContainerLow,
-    alignSelf: 'center',
-  },
-  arrowBtnDisabled: {
-    opacity: 0.3,
-  },
-  itemCard: {
-    flex: 1,
-    borderRadius: Radius.lg,
-    overflow: 'hidden',
-    backgroundColor: Palette.surfaceContainerLow,
-  },
-  itemImage: {
-    width: '100%',
-    height: '100%',
-  },
-  itemPlaceholder: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    backgroundColor: Palette.surfaceContainerLow,
-    borderRadius: Radius.lg,
-  },
-  placeholderText: {
-    ...Typography.bodyMd,
-    color: Palette.onSurfaceVariant,
-  },
-  counter: {
-    ...Typography.labelSm,
-    color: Palette.onSurfaceVariant,
-    textAlign: 'center',
   },
   actions: {
     paddingVertical: Spacing.stackMd,
@@ -421,97 +210,8 @@ const styles = StyleSheet.create({
     ...Typography.titleLg,
     color: Palette.onPrimary,
   },
-  modalOverlay: {
-    flex: 1,
-    backgroundColor: 'rgba(0,0,0,0.4)',
-    justifyContent: 'flex-end',
-  },
-  modalSheet: {
-    backgroundColor: Palette.surfaceContainerLowest,
-    borderTopLeftRadius: Radius.lg,
-    borderTopRightRadius: Radius.lg,
-    padding: Spacing.containerMargin,
-    gap: Spacing.stackMd,
-  },
-  modalHandle: {
-    width: 40,
-    height: 4,
-    borderRadius: Radius.full,
-    backgroundColor: Palette.outlineVariant,
-    alignSelf: 'center',
-    marginBottom: Spacing.stackSm,
-  },
-  modalTitle: {
-    ...Typography.headlineMd,
-    color: Palette.onSurface,
-  },
-  nameInput: {
-    ...Typography.bodyMd,
-    color: Palette.onSurface,
-    borderWidth: 1,
-    borderColor: Palette.outlineVariant,
-    borderRadius: Radius.md,
-    paddingHorizontal: Spacing.stackMd,
-    paddingVertical: Spacing.stackSm,
-  },
-  favRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: Spacing.stackMd,
-    paddingVertical: Spacing.stackSm,
-  },
-  favLabel: {
-    ...Typography.bodyMd,
-    color: Palette.onSurface,
-  },
-  collectionsRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    paddingVertical: Spacing.stackSm,
-    borderTopWidth: 1,
-    borderTopColor: Palette.outlineVariant,
-  },
-  collectionsLabel: {
-    ...Typography.bodyMd,
-    color: Palette.onSurface,
-  },
-  comingSoon: {
-    ...Typography.labelSm,
-    color: Palette.onSurfaceVariant,
-  },
   errorText: {
     ...Typography.bodyMd,
     color: Palette.error,
-  },
-  modalActions: {
-    flexDirection: 'row',
-    gap: Spacing.stackMd,
-    paddingTop: Spacing.stackSm,
-  },
-  cancelBtn: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingVertical: Spacing.stackMd,
-    borderRadius: Radius.md,
-    borderWidth: 1,
-    borderColor: Palette.outline,
-  },
-  cancelBtnText: {
-    ...Typography.titleLg,
-    color: Palette.onSurface,
-  },
-  confirmBtn: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingVertical: Spacing.stackMd,
-    borderRadius: Radius.md,
-    backgroundColor: Palette.primary,
-  },
-  confirmBtnText: {
-    ...Typography.titleLg,
-    color: Palette.onPrimary,
   },
 });
